+++ Webapp/src/gen/java/io/swagger/api/MortgagecalculationApiService.java
@@ -1,23 +1,10 @@
 package io.swagger.api;
 
-import io.swagger.api.*;
-import io.swagger.model.*;
-
-import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
-
-import io.swagger.model.InlineResponse2001;
-import io.swagger.model.InlineResponse4001;
 import io.swagger.model.MortgageCalculator;
-
-import java.util.Map;
-import java.util.List;
-import io.swagger.api.NotFoundException;
-
-import java.io.InputStream;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.SecurityContext;
-import javax.validation.constraints.*;
+
 @javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJerseyServerCodegen", date = "2021-01-25T01:46:26.939Z[GMT]")public abstract class MortgagecalculationApiService {
     public abstract Response mortgagecalculationPost(MortgageCalculator body,SecurityContext securityContext) throws NotFoundException;
 }
