+++ Webapp/src/gen/java/io/swagger/model/Cardnumber.java
@@ -12,12 +12,10 @@
 
 package io.swagger.model;
 
-import java.util.Objects;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
 import io.swagger.v3.oas.annotations.media.Schema;
-import javax.validation.constraints.*;
-import javax.validation.Valid;
+
+import java.util.Objects;
 
 /**
  * Cardnumber
